Refactor: make changes according to comments
